requirements.txt
streamlit
PyPDF2
python-docx
googletrans==4.0.0-rc1
